refactor(hooks): colocate single-area hook groups (wave 2)
- hooks/dependencies  -> modules/MainApp/Integrations/hooks (useSystemDependencies,
                          usePrerequisiteCheck) and app/root/usePostPaintGitProbe.ts
- hooks/git/sourceControl -> modules/WorkStation/CodeEditor/hooks/sourceControl;
                          generateCommitMessage (a bare invoke wrapper) -> api/tauri/git/commitMessage.ts
                          so the ChatPanel test does not import from modules/
- hooks/testRunner    -> modules/WorkStation/CodeEditor/hooks/useTestRunner.ts
- hooks/benchmark     -> features/BenchmarkPanel/hooks (WorkStation sidebar imports the feature)

Group barrels removed; consumers import files directly.
