--- src/hooks/dependencies/index.ts
@@ -1,2 +1,0 @@
-export * from "./useSystemDependencies";
-export * from "./usePrerequisiteCheck";
