--- src/hooks/testRunner/index.ts
@@ -1,8 +1,0 @@
-/**
- * Test Runner Hooks Exports
- */
-export {
-  useTestRunner,
-  type UseTestRunnerOptions,
-  type UseTestRunnerReturn,
-} from "./useTestRunner";
